Uses the UTC timezone for the dates
Magento saves most dates (including the quote's updated_at field) in the UTC timezone. Therefore any check on the quote's updated_at field must use values in the same timezone, otherwise wrong results will be obtained.

I tried to replace the call to date() with a call to utcdate(), but that didn't seem to work. That's why in the PR I set the timezone after the fact.