Deletes followups that are no longer valid
After a followup object has been created it can happen that the quote that it references gets updated or even converted into an order. If that happens it is not correct to keep on sending alerts about a cart that is obviously no longer abandoned.

This PR deletes any such invalid followup to stop the sending of these inappropriate emails.