Fixes an off-by-one error in the handling of the delay configurations
After sending the first email the value of the "offset" field is 0. The _getNextDate() implementation then checks the delays with offset 0 and 1, but the configuration delay fields have offset 1, 2 and 3. As such $delayCurrent will be set to null, which causes the $delayReal value to be incorrect. Similarly, after sending the second email the delays will again be wrong, causing a wrong value for the $delayReal variable.

This fix changes the lines that read the config so that the correct values are retrieved. In addition there is a check not to update the mail_scheduled_at field after sending the last mail, as that would again access a nonexistent configuration field.